Add method for creating a new entry
Starts the timer
