+++ src/start/start.py
@@ -1,7 +1,10 @@
 import click
 
+from ..csv_handler import CSVHandler
+
 
 @click.command()
 def start():
     """Start a new timer"""
-    pass
+    csv_handler = CSVHandler.CSVHandler()
+    csv_handler.create_new_entry()
